+++ sdk/mixedreality/azure-resourcemanager-mixedreality/src/main/java/com/azure/resourcemanager/mixedreality/fluent/MixedRealityClient.java
@@ -61,13 +61,6 @@ public interface MixedRealityClient {
     ResourceProvidersClient getResourceProviders();
 
     /**
-     * Gets the SpatialAnchorsAccountsClient object to access its operations.
-     * 
-     * @return the SpatialAnchorsAccountsClient object.
-     */
-    SpatialAnchorsAccountsClient getSpatialAnchorsAccounts();
-
-    /**
      * Gets the RemoteRenderingAccountsClient object to access its operations.
      * 
      * @return the RemoteRenderingAccountsClient object.