+++ sdk/mixedreality/azure-resourcemanager-mixedreality/CHANGELOG.md
@@ -1,14 +1,28 @@
 # Release History
 
-## 1.1.0-beta.1 (Unreleased)
+## 1.0.0-beta.1 (2025-01-15)
 
-### Features Added
+- Azure Resource Manager MixedReality client library for Java. This package contains Microsoft Azure SDK for MixedReality Management SDK. Mixed Reality Client. Package tag package-2021-03-01-preview. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Breaking Changes
 
-### Bugs Fixed
+#### `models.SpatialAnchorsAccount$DefinitionStages` was removed
 
-### Other Changes
+#### `models.SpatialAnchorsAccount` was removed
+
+#### `models.SpatialAnchorsAccountPage` was removed
+
+#### `models.SpatialAnchorsAccount$Update` was removed
+
+#### `models.SpatialAnchorsAccount$Definition` was removed
+
+#### `models.SpatialAnchorsAccounts` was removed
+
+#### `models.SpatialAnchorsAccount$UpdateStages` was removed
+
+#### `MixedRealityManager` was modified
+
+* `spatialAnchorsAccounts()` was removed
 
 ## 1.0.0 (2024-12-23)